Revert "Eject React application"
This reverts commit 7891d7fe230e71e536fd0d6d31753cf6c6b854ae.
