--- config/jest/babelTransform.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const babelJest = require('babel-jest');
-
-module.exports = babelJest.createTransformer({
-  presets: [require.resolve('babel-preset-react-app')],
-  
-});
