--- config/jest/graphqlTransform.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const loader = require('graphql-tag/loader');
-
-module.exports = {
-  process(src) {
-    return loader.call({ cacheable() {} }, src);
-  },
-};
