+++ package.json
@@ -3,8 +3,6 @@
   "version": "2.4.12",
   "private": true,
   "dependencies": {
-    "@babel/core": "7.0.0-beta.46",
-    "@babel/runtime": "7.0.0-beta.46",
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-svg-core": "^1.2.0-14",
     "@fortawesome/free-brands-svg-icons": "^5.1",
@@ -12,92 +10,47 @@
     "@fortawesome/free-solid-svg-icons": "^5.1",
     "@fortawesome/react-fontawesome": "^0.1",
     "@sentry/browser": "^5.15.4",
-    "autoprefixer": "8.5.0",
     "axios": "^0.19.2",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "8.2.3",
-    "babel-jest": "22.4.3",
-    "babel-loader": "8.0.0-beta.0",
-    "babel-plugin-named-asset-import": "1.0.0-next.3e165448",
-    "babel-preset-react-app": "4.0.0-next.3e165448",
-    "bfj": "5.2.0",
+    "babel-eslint": "^10.0.1",
     "bootswatch": "^4.1.1",
-    "case-sensitive-paths-webpack-plugin": "2.1.2",
-    "chalk": "2.4.1",
     "concurrently": "^4.0.0",
-    "css-loader": "0.28.11",
     "date-fns": "^1.29.0",
-    "dotenv": "5.0.1",
-    "dotenv-expand": "4.2.0",
     "dynamic-cdn-webpack-plugin": "^4.0.0-rc.1",
-    "eslint": "4.19.1",
-    "eslint-config-react-app": "3.0.0-next.3e165448",
-    "eslint-loader": "2.0.0",
-    "eslint-plugin-flowtype": "2.46.3",
-    "eslint-plugin-import": "2.11.0",
-    "eslint-plugin-jsx-a11y": "6.0.3",
-    "eslint-plugin-react": "7.8.2",
-    "file-loader": "1.1.11",
-    "fs-extra": "5.0.0",
-    "graphql": "0.13.2",
-    "graphql-tag": "2.9.2",
-    "html-webpack-plugin": "3.2.0",
+    "eslint-plugin-react": "^7.11.1",
     "husky": "^1.1.1",
     "i18next": "^14.0.1",
     "i18next-browser-languagedetector": "^3.0.1",
     "i18next-xhr-backend": "^2.0.1",
-    "identity-obj-proxy": "3.0.0",
-    "jest": "22.4.3",
     "lint-staged": "^7.3.0",
-    "loader-utils": "^1.1.0",
     "lodash": "^4.17.13",
     "md5": "^2.2.1",
-    "mini-css-extract-plugin": "^0.4.0",
     "module-to-cdn": "^3.1.2",
-    "object-assign": "4.1.1",
-    "optimize-css-assets-webpack-plugin": "^4.0.1",
-    "postcss-flexbugs-fixes": "3.3.1",
-    "postcss-loader": "2.1.5",
-    "promise": "8.0.1",
     "prop-types": "^15.6.2",
     "qs": "^6.5.2",
-    "raf": "3.4.0",
-    "react": "^16.13.0",
+    "react": "^16.13.1",
     "react-cache": "^2.0.0-alpha.0",
     "react-day-picker": "^7.1.10",
-    "react-dev-utils": "6.0.0-next.3e165448",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.13.1",
     "react-ga": "^2.5.3",
     "react-i18next": "^10.6.1",
     "react-lazy-load-image-component": "^1.3.2",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
+    "react-scripts": "3.4.1",
     "reactstrap": "6.5.0",
     "redux": "^4.0.0",
     "redux-debounced": "^0.5.0",
     "redux-logger": "^3.0.6",
     "redux-promise-middleware": "^5.1.1",
-    "resolve": "1.6.0",
-    "sass-loader": "7.0.1",
     "shortid": "^2.2.8",
-    "style-loader": "0.21.0",
-    "svgr": "1.9.2",
-    "sw-precache-webpack-plugin": "0.11.5",
-    "thread-loader": "1.1.5",
-    "uglifyjs-webpack-plugin": "1.2.5",
-    "url-loader": "1.0.1",
-    "webpack": "4.8.3",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-dev-server": "3.1.11",
-    "webpack-manifest-plugin": "2.0.3",
-    "whatwg-fetch": "2.0.4"
+    "webpack-bundle-analyzer": "^3.3.2"
   },
   "scripts": {
-    "start": "concurrently -r --kill-others 'node scripts/start.js' 'docker-compose up'",
-    "build": "REACT_APP_VERSION=$npm_package_version node scripts/build.js",
+    "start": "concurrently -r --kill-others 'react-scripts start' 'docker-compose up'",
+    "build": "REACT_APP_VERSION=$npm_package_version react-scripts build",
     "lint": "node_modules/.bin/eslint -c .eslintrc src/**/*.js",
-    "test": "node scripts/test.js --env=jsdom",
-    "patch": "yarn version --no-git-tag-version --patch"
+    "patch": "yarn version --no-git-tag-version --patch",
+    "eject": "react-scripts eject"
   },
   "husky": {
     "hooks": {
@@ -124,45 +77,5 @@
       "ie 11"
     ]
   },
-  "proxy": "http://localhost:3274",
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx,mjs}"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testMatch": [
-      "**/__tests__/**/*.{js,jsx,mjs}",
-      "**/?(*.)(spec|test).{js,jsx,mjs}"
-    ],
-    "roots": [
-      "<rootDir>/src"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.(js|jsx|mjs)$": "<rootDir>/config/jest/babelTransform.js",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^.+\\.(graphql)$": "<rootDir>/config/jest/graphqlTransform.js",
-      "^(?!.*\\.(js|jsx|mjs|css|json|graphql)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$",
-      "^.+\\.module\\.(css|sass|scss)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web",
-      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
-    },
-    "moduleFileExtensions": [
-      "web.js",
-      "js",
-      "json",
-      "web.jsx",
-      "jsx",
-      "node",
-      "mjs"
-    ]
-  }
+  "proxy": "http://localhost:3274"
 }
