+++ src/registerServiceWorker.js
@@ -9,6 +9,7 @@
 // This link also includes instructions on opting out of this behavior.
 
 /* eslint-disable no-console */
+
 import {
   NEW_VERSION_READY,
 } from 'Constants';
